Skip Docker test when Docker is not available
Use a conditional check for Docker during test initialization.

Similar to the technique used in docker-plugin and elsewhere so that
the tests do not fail when Docker is unavailable.

https://github.com/jenkinsci/eddsa-api-plugin/commit/dcf1d647a934603bb5bf0611684768055853b3a0#r153984874

https://github.com/jenkinsci/bom/pull/4739#pullrequestreview-2699468137

Testing done:

Confirmed the tests pass on Ubuntu 22.04 with Java 21 and Docker available
both before and after this change.

Confirmed EddsaTest fails before this change on RHEL 8 with Java 21 and
no Docker available.

Confirmed EddsaTest passes after this change on RHEL 8 with Java 21 and
no Docker available.